setup the ddp algorithm for the LIPM and started setting up a visualizer for it

+++ ihmc-common-walking-control-modules/src/main/java/us/ihmc/commonWalkingControlModules/dynamicPlanning/DiscreteHybridDynamics.java
@@ -5,8 +5,8 @@ import org.ejml.data.DenseMatrix64F;
 public interface DiscreteHybridDynamics<E extends Enum>
 {
    void setTimeStepSize(double deltaT);
-   double getStateVectorSize();
-   double getControlVectorSize();
+   int getStateVectorSize();
+   int getControlVectorSize();
 
    /** f */
    void getDynamics(E hybridState, DenseMatrix64F currentState, DenseMatrix64F currentControl, DenseMatrix64F matrixToPack);
